turns out that there was a bug
I'm not sure what changed, but when I tried running this today, the
b.CurrentPrivateWorkingSet was stale, so setting startSet failed. This
was working before. Perhaps something about the way the CookiesAndCache
are cleared caused an issue (calling that process caused MicrosoftEdge
to fall out of scope... Changing slightly to account for that.